+++ src/UI/index.ts
@@ -24,6 +24,7 @@ export * from './walletConnect/WalletConnectLoginButton';
 export * from './walletConnect/WalletConnectLoginContainer';
 export * from './webWallet/WebWalletLoginButton';
 export * from './webWallet/XaliasLoginButton';
+export * from './webWallet/CrossWindowLoginButton';
 export * from './TransactionsTable';
 export * from './ScamPhishingAlert';
 export * from './TimeAgo';