+++ src/UI/walletConnect/WalletConnectLoginContainer/WalletConnectLoginContent/WalletConnectLoginContent.tsx
@@ -98,12 +98,8 @@ const WalletConnectLoginContentComponent = ({
   }, [walletConnectUriV2]);
 
   useEffect(() => {
-    if (canLoginRef?.current === false) {
-      return;
-    }
-
     initLoginWithWalletConnectV2();
-  }, [canLoginRef?.current]);
+  }, []);
 
   const authorizationInfo = showScamPhishingAlert
     ? getAuthorizationInfo(token, containerScamPhishingAlertClassName)