+++ src/hooks/account/useGetIsWalletConnectV2Initialized.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'reduxStore/DappProviderContext';
+import { isWalletConnectV2InitializedSelector } from 'reduxStore/selectors';
+
+export const useGetIsWalletConnectV2Initialized = () => {
+  return useSelector(isWalletConnectV2InitializedSelector);
+};