+++ src/hooks/login/useWalletConnectV2Login.ts
@@ -6,9 +6,8 @@ import { setAccountProvider } from 'providers/accountProvider';
 import { emptyProvider, getProviderType } from 'providers/utils';
 import { loginAction } from 'reduxStore/commonActions';
 import { useDispatch, useSelector } from 'reduxStore/DappProviderContext';
-import { logoutRouteSelector } from 'reduxStore/selectors';
+import { loginMethodSelector, logoutRouteSelector } from 'reduxStore/selectors';
 import {
-  chainIDSelector,
   walletConnectDeepLinkSelector,
   walletConnectV2OptionsSelector,
   walletConnectV2ProjectIdSelector,
@@ -16,13 +15,11 @@ import {
 } from 'reduxStore/selectors/networkConfigSelectors';
 import { setWalletConnectLogin } from 'reduxStore/slices';
 import { LoginMethodsEnum } from 'types/enums.types';
-import {
-  LoginHookGenericStateType,
-  OnProviderLoginType
-} from 'types/login.types';
+import { getHasNativeAuth } from 'utils/getHasNativeAuth';
 import { getIsLoggedIn } from 'utils/getIsLoggedIn';
 import { optionalRedirect } from 'utils/internal';
 import { logout } from 'utils/logout';
+import { waitForChainID } from 'utils/waitForChainID';
 import {
   PairingTypes,
   SessionEventTypes,
@@ -30,6 +27,10 @@ import {
   WalletConnectV2Provider
 } from 'utils/walletconnect/__sdkWalletconnectProvider';
 import { getWindowLocation } from 'utils/window/getWindowLocation';
+import {
+  LoginHookGenericStateType,
+  OnProviderLoginType
+} from '../../types/login.types';
 import { useLoginService } from './useLoginService';
 
 export enum WalletConnectV2Error {
@@ -40,7 +41,8 @@ export enum WalletConnectV2Error {
   connectError = 'Unable to connect',
   userRejected = 'User rejected connection proposal',
   userRejectedExisting = 'User rejected existing connection proposal',
-  errorLogout = 'Unable to remove existing pairing'
+  errorLogout = 'Unable to remove existing pairing',
+  invalidChainID = 'Invalid chainID'
 }
 
 export interface InitWalletConnectV2Type extends OnProviderLoginType {
@@ -59,7 +61,7 @@ export interface WalletConnectV2LoginHookCustomStateType {
 }
 
 export type WalletConnectV2LoginHookReturnType = [
-  (loginProvider?: boolean) => void,
+  (loginProvider?: boolean) => Promise<void>,
   LoginHookGenericStateType,
   WalletConnectV2LoginHookCustomStateType
 ];
@@ -74,8 +76,8 @@ export const useWalletConnectV2Login = ({
   customRequestMethods = []
 }: InitWalletConnectV2Type): WalletConnectV2LoginHookReturnType => {
   const dispatch = useDispatch();
-  const hasNativeAuth = nativeAuth != null;
-  const loginService = useLoginService(nativeAuth);
+  const hasNativeAuth = getHasNativeAuth(nativeAuth);
+  const loginService = useLoginService(hasNativeAuth ? nativeAuth : false);
   let token = tokenToSign;
 
   const [error, setError] = useState<string>('');
@@ -84,19 +86,22 @@ export const useWalletConnectV2Login = ({
   const [wcPairings, setWcPairings] = useState<
     PairingTypes.Struct[] | undefined
   >([]);
+  const [sessionProvider, setSessionProvider] =
+    useState<WalletConnectV2Provider | null>(null);
 
-  const { provider, providerType } = useGetAccountProvider();
+  const { provider } = useGetAccountProvider();
   const walletConnectV2RelayAddress = useSelector(walletConnectV2RelaySelector);
   const walletConnectV2ProjectId = useSelector(
     walletConnectV2ProjectIdSelector
   );
   const walletConnectV2Options = useSelector(walletConnectV2OptionsSelector);
-  const chainId = useSelector(chainIDSelector);
   const walletConnectDeepLink = useSelector(walletConnectDeepLinkSelector);
   const dappLogoutRoute = useSelector(logoutRouteSelector);
+  const loginMethod = useSelector(loginMethodSelector);
   const providerRef = useRef<any>(provider);
   const canLoginRef = parentCanLoginRef ?? useRef<boolean>(false);
   const isInitialisingRef = useRef<boolean>(false);
+  const mounted = useRef(false);
 
   const logoutRoute = providerLogoutRoute ?? dappLogoutRoute ?? '/';
   const dappMethods: string[] = [
@@ -129,17 +134,7 @@ export const useWalletConnectV2Login = ({
 
   const handleOnLogin = async () => {
     try {
-      const isLoggedIn = getIsLoggedIn();
-      const isLoggedInWithDifferentProvider =
-        isLoggedIn &&
-        providerType &&
-        providerType !== LoginMethodsEnum.walletconnectv2;
-
-      if (
-        isInitialisingRef.current ||
-        providerRef.current == null ||
-        isLoggedInWithDifferentProvider
-      ) {
+      if (isInitialisingRef.current || providerRef.current == null) {
         return;
       }
 
@@ -210,7 +205,7 @@ export const useWalletConnectV2Login = ({
       }
 
       providerRef.current = emptyProvider;
-      setAccountProvider(emptyProvider);
+      setSessionProvider(null);
     } catch {
       console.warn('Unable to logout');
     }
@@ -281,26 +276,24 @@ export const useWalletConnectV2Login = ({
   };
 
   async function initiateLogin(loginProvider = true) {
+    const chainId = await waitForChainID({ maxRetries: 15 });
+
+    if (!chainId) {
+      setError(WalletConnectV2Error.invalidChainID);
+      return;
+    }
+
     if (!walletConnectV2ProjectId || !walletConnectV2RelayAddress) {
       setError(WalletConnectV2Error.invalidConfig);
       return;
     }
 
     const isLoggedIn = getIsLoggedIn();
-    const isLoggedInWithDifferentProvider =
-      isLoggedIn &&
-      providerType &&
-      providerType !== LoginMethodsEnum.walletconnectv2;
 
     const cannotLogin = canLoginRef.current === false && !isLoggedIn;
     const isInitialized = providerRef.current?.isInitialized?.();
 
-    if (
-      isInitialisingRef.current ||
-      cannotLogin ||
-      isLoggedInWithDifferentProvider ||
-      isInitialized
-    ) {
+    if (isInitialisingRef.current || cannotLogin || isInitialized) {
       return;
     }
 
@@ -309,7 +302,7 @@ export const useWalletConnectV2Login = ({
     if (providerRef.current?.walletConnector) {
       providerRef.current.init();
 
-      setAccountProvider(providerRef.current);
+      setSessionProvider(providerRef.current);
 
       isInitialisingRef.current = false;
       canLoginRef.current = true;
@@ -336,7 +329,7 @@ export const useWalletConnectV2Login = ({
     );
 
     await newProvider.init();
-    setAccountProvider(newProvider);
+    setSessionProvider(newProvider);
     providerRef.current = newProvider;
     isInitialisingRef.current = false;
     canLoginRef.current = true;
@@ -359,6 +352,11 @@ export const useWalletConnectV2Login = ({
       const { uri, approval } = await providerRef.current?.connect({
         methods: dappMethods
       });
+
+      // Do not do any other actions if component is not mounted
+      if (!mounted.current) {
+        return;
+      }
 
       const hasUri = Boolean(uri);
 
@@ -407,18 +405,36 @@ export const useWalletConnectV2Login = ({
   }, [tokenToSign, providerRef.current?.connect]);
 
   useUpdateEffect(() => {
-    if (canLoginRef?.current === false) {
-      cancelLogin();
-    }
-  }, [canLoginRef.current]);
-
-  useUpdateEffect(() => {
     providerRef.current = provider;
   }, [provider]);
 
   useEffect(() => {
+    mounted.current = true;
+
+    return () => {
+      mounted.current = false;
+    };
+  }, []);
+
+  useEffect(() => {
     setIsLoading(!Boolean(wcUri));
   }, [wcUri]);
+
+  useEffect(() => {
+    if (!sessionProvider) {
+      return;
+    }
+
+    // Check if a new session has been created is already connected
+    const isConnected =
+      Boolean(sessionProvider.session) ||
+      loginMethod === LoginMethodsEnum.walletconnectv2;
+
+    // Set new provider only if account is logged in and if walletConnect session is available
+    if (isConnected && isLoggedIn) {
+      setAccountProvider(sessionProvider);
+    }
+  }, [sessionProvider, isLoggedIn]);
 
   return [
     initiateLogin,